+++ Changelog.md
@@ -2,7 +2,7 @@
 
 ## v3.5.3
 
-- Manual Dependencies: Linux Dependencies (`rpm`, `apk`, `deb`) can be provided as reference dependency in fossa-deps file []().
+- Manual Dependencies: Linux Dependencies (`rpm-generic`, `apk`, `deb`) can be provided as reference dependency in fossa-deps file [#1086](https://github.com/fossas/fossa-cli/pull/1086).
 
 ## v3.5.2